Update getting-started.mdx
The `emitDecoratorMetadata` option was set to `false` here can't cause the DI of **tsyringe** do not work later. When a `injectable()` class with constructor injection was resolved, it would always throw an exception. But in the [tsconfig.json](https://github.com/discordx-ts/discordx/blob/main/packages/discordx/examples/di/tsyringe/tsconfig.json) of the example for **tsyringe**, the `emitDecoratorMetadata` option was set to `true`, and it works. 
The document does not let people know that changing this option when using DI will make them stuck at this exception. I hope this helps!